FIX+MAINT fix delete_text_at_span, move to stac.util.doc
